♻️ 1GEWE2 blueprints: decompose model types
