+++ .agentplane/tasks/202605290030-1GEWE2/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290030-1GEWE2`
+Title: Blueprint model decomposition
+Canonical task record: `.agentplane/tasks/202605290030-1GEWE2/README.md`
+
+## Summary
+
+Blueprint model decomposition
+
+Decompose packages/agentplane/src/blueprints/model.ts into focused blueprint model schema/type modules while preserving public blueprint model exports and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/blueprints/model.ts into focused blueprint model schema/type modules while preserving public blueprint model exports and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "Blueprint model decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T00:30:51.886Z
+- Branch: task/202605290030-1GEWE2/blueprint-model-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
