+++ .agentplane/tasks/202605290030-1GEWE2/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 1GEWE2 task: Blueprint model decomposition [202605290030-1GEWE2]
