+++ .agentplane/tasks/202605290030-1GEWE2/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T00:30:51.886Z
+
+## Task
+
+- Task: `202605290030-1GEWE2`
+- Title: Blueprint model decomposition
+- Status: DOING
+- Branch: `task/202605290030-1GEWE2/blueprint-model-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290030-1GEWE2/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T00:30:51.886Z
+- Branch: task/202605290030-1GEWE2/blueprint-model-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
